extract css styles to be closer to the component that uses that
