+++ application/ui/src/features/remote-trainers/remote-trainers-page.module.css
@@ -1,20 +1,7 @@
-.table {
+.container {
     border: 1px solid var(--spectrum-global-color-gray-300);
     border-radius: var(--spectrum-global-dimension-size-75);
     overflow: hidden;
-}
-
-.tableHeader {
-    padding: var(--spectrum-global-dimension-size-150) var(--spectrum-global-dimension-size-200);
-    background: var(--spectrum-global-color-gray-200);
-    text-transform: uppercase;
-    font-size: var(--spectrum-global-dimension-font-size-50);
-}
-
-.expandedContent {
-    padding: var(--spectrum-global-dimension-size-300);
-    background: var(--spectrum-global-color-gray-75);
-    border-top: 1px solid var(--spectrum-global-color-gray-300);
 }
 
 .addButton {
@@ -33,121 +20,8 @@
     padding-block: var(--spectrum-global-dimension-size-50);
 }
 
-.trainerName {
-    font-weight: 700;
-    font-size: var(--spectrum-global-dimension-font-size-200);
-}
-
-.url {
-    overflow-wrap: anywhere;
-    color: var(--spectrum-global-color-gray-700);
-}
-
-.cardMeta {
-    margin-top: var(--spectrum-global-dimension-size-75);
-}
-
-.cardMetaText {
-    color: var(--spectrum-global-color-gray-600);
-    font-size: var(--spectrum-global-dimension-font-size-50);
-}
-
-.detailSection {
-    padding: var(--spectrum-global-dimension-size-200);
-    background: var(--spectrum-global-color-gray-100);
-    border: 1px solid var(--spectrum-global-color-gray-300);
-    border-radius: var(--spectrum-global-dimension-size-75);
-}
-
-.sectionHeading {
-    margin-top: 0;
-    color: var(--spectrum-global-color-gray-700);
-    font-size: var(--spectrum-global-dimension-font-size-100);
-    letter-spacing: 0.06em;
-    text-transform: uppercase;
-}
-
-.checkList {
-    display: grid;
-    grid-template-columns: 1.375rem minmax(0, 1fr) max-content;
-    column-gap: var(--spectrum-global-dimension-size-100);
-    margin-top: var(--spectrum-global-dimension-size-100);
-}
-
-.checkRow {
-    border-bottom: 1px solid var(--spectrum-global-color-gray-200);
-}
-
-.checkRow:last-child {
-    border-bottom: 0;
-}
-
-.checkIcon {
-    display: grid;
-    place-items: center;
-    line-height: 1;
-    font-weight: 700;
-}
-
-.positive {
-    color: var(--spectrum-global-color-static-green-500);
-}
-
-.yellow {
-    color: var(--spectrum-global-color-static-orange-500);
-}
-
-.negative {
-    color: var(--spectrum-global-color-static-red-500);
-}
-
-.neutral {
-    color: var(--spectrum-global-color-gray-600);
-}
-
-.checkLabel {
-    font-weight: 500;
-}
-
-.checkContent {
-    display: flex;
-    align-items: baseline;
-    gap: var(--spectrum-global-dimension-size-75);
-    min-width: 0;
-}
-
-.checkDetail {
-    color: var(--spectrum-global-color-gray-700);
-    font-size: var(--spectrum-global-dimension-font-size-75);
-}
-
-.checkRow > :last-child {
-    justify-self: end;
-    white-space: nowrap;
-}
-
-.definitionList {
-    display: grid;
-    grid-template-columns: minmax(9rem, max-content) minmax(0, 1fr);
-    gap: var(--spectrum-global-dimension-size-100) var(--spectrum-global-dimension-size-200);
-    margin: var(--spectrum-global-dimension-size-150) 0 0;
-}
-
-.definitionList dt {
-    color: var(--spectrum-global-color-gray-700);
-}
-
-.definitionList dd {
-    margin: 0;
-    overflow-wrap: anywhere;
-}
-
 .emptyList {
     display: block;
     padding: var(--spectrum-global-dimension-size-150);
     color: var(--spectrum-global-color-gray-700);
-}
-
-.errorMessage {
-    color: var(--spectrum-global-color-static-red-600);
 }
