+++ application/ui/src/features/remote-trainers/remote-trainers-table/remote-trainer-detail/remote-trainer-detail.module.css
@@ -1,0 +1,77 @@
+.sectionHeading {
+    margin-top: 0;
+    color: var(--spectrum-global-color-gray-700);
+    font-size: var(--spectrum-global-dimension-font-size-100);
+    letter-spacing: 0.06em;
+    text-transform: uppercase;
+}
+
+.definitionList {
+    display: grid;
+    grid-template-columns: minmax(9rem, max-content) minmax(0, 1fr);
+    gap: var(--spectrum-global-dimension-size-100) var(--spectrum-global-dimension-size-200);
+    margin: var(--spectrum-global-dimension-size-150) 0 0;
+}
+
+.definitionList dt {
+    color: var(--spectrum-global-color-gray-700);
+}
+
+.checkList {
+    display: grid;
+    grid-template-columns: 1.375rem minmax(0, 1fr) max-content;
+    column-gap: var(--spectrum-global-dimension-size-100);
+    margin-top: var(--spectrum-global-dimension-size-100);
+}
+
+.checkContent {
+    display: flex;
+    align-items: baseline;
+    gap: var(--spectrum-global-dimension-size-75);
+    min-width: 0;
+}
+
+.checkLabel {
+    font-weight: 500;
+}
+
+.checkDetail {
+    color: var(--spectrum-global-color-gray-700);
+    font-size: var(--spectrum-global-dimension-font-size-75);
+}
+
+.checkRow > :last-child {
+    justify-self: end;
+    white-space: nowrap;
+}
+
+.checkRow {
+    border-bottom: 1px solid var(--spectrum-global-color-gray-200);
+}
+
+.checkRow:last-child {
+    border-bottom: 0;
+}
+
+.checkIcon {
+    display: grid;
+    place-items: center;
+    line-height: 1;
+    font-weight: 700;
+}
+
+.positive {
+    color: var(--spectrum-global-color-static-green-500);
+}
+
+.yellow {
+    color: var(--spectrum-global-color-static-orange-500);
+}
+
+.negative {
+    color: var(--spectrum-global-color-static-red-500);
+}
+
+.neutral {
+    color: var(--spectrum-global-color-gray-600);
+}
