+++ application/ui/src/features/remote-trainers/remote-trainers-table/remote-trainer-form/remote-trainer-form.module.css
@@ -1,0 +1,3 @@
+.errorMessage {
+    color: var(--spectrum-global-color-static-red-600);
+}
